agents/ten_packages/extension/tavilysearch_tool_python/requirements.txt
aiohttp
tavily-python>=0.1.0